+++ oap-server/server-alarm-plugin/src/test/java/org/apache/skywalking/oap/server/core/alarm/provider/RulesReaderTest.java
@@ -74,8 +74,10 @@ public class RulesReaderTest {
         WebhookSettings rulesWebhooks = rules.getWebhookSettingsMap().get(AlarmHooksType.webhook.name() + ".default");
         Assertions.assertEquals(2, rulesWebhooks.getUrls().size());
         Assertions.assertEquals("http://127.0.0.1/go-wechat/", rulesWebhooks.getUrls().get(1));
-        Assertions.assertEquals("Bearer", rulesWebhooks.getAuthorization().getType().getName());
-        Assertions.assertEquals("dummy_token", rulesWebhooks.getAuthorization().getCredentials());
+        WebhookSettings rulesWebhooks2 = rules.getWebhookSettingsMap().get(AlarmHooksType.webhook.name() + ".custom1");
+        Assertions.assertEquals(2, rulesWebhooks2.getHeaders().size());
+        Assertions.assertEquals("Bearer bearer_token", rulesWebhooks2.getHeaders().get("Authorization"));
+        Assertions.assertEquals("arbitrary-additional-http-headers", rulesWebhooks2.getHeaders().get("x-company-header"));
 
         GRPCAlarmSetting grpcAlarmSetting = rules.getGrpcAlarmSettingMap().get(AlarmHooksType.gRPC.name() + ".default");
         assertNotNull(grpcAlarmSetting);